mccarthy/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
cffi==1.12.3
Click==7.0
cryptography==2.1.4
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
opencv-python==4.1.1.26
pandas==0.25.0
Pillow==6.1.0
protobuf==3.8.0
pycparser==2.19
pycrypto==2.6.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
pyxdg==0.25
PyYAML==5.1.2
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
SecretStorage==2.3.1
six==1.11.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2
